Allow for large scale factors.  C99 warrants that ULLONG_MAX is not less than 2^64-1; and uintmax_t cannot be more narrow than unsigned long long. This allows for scale factors up to Exa inclusively.
Use plain int for the scale index to be consistent
with ifcmds.c and enum.

+++ usr.bin/systat/convtbl.h
@@ -32,10 +32,11 @@
 #define _CONVTBL_H_
 
 #include <sys/types.h>
+#include <stdint.h>
 
 #define BITS	(1)
 #define BYTES	(1)
-#define KILO	(1024)
+#define KILO	(1024LL)
 #define	MEGA	(KILO * 1024)
 #define GIGA	(MEGA * 1024)
 
@@ -55,14 +56,14 @@ enum scale {
 #define BYTE	(1)
 
 struct convtbl {
-	u_int	 mul;
-	u_int	 scale;
+	uintmax_t	 mul;
+	uintmax_t	 scale;
 	const char	*str;
 };
 
 extern	struct convtbl convtbl[];
 
-extern	double 	 convert(const u_long, const u_int);
-extern	const char	*get_string(const u_long, const u_int);
+extern	double		 convert(const uintmax_t, const int);
+extern	const char	*get_string(const uintmax_t, const int);
 
 #endif		/* ! _CONVTBL_H_ */
